style: minor code cleanups
- drop redundant @return docblocks and an (int) cast
- ExternalXtreamService::resolvePublicIps → instance method
- invert empty-else guard in ListingsController
- gitignore .cache/ and coverage/

+++ src/Core/Util/ImageResizeService.php
@@ -345,7 +345,7 @@ class ImageResizeService {
 		$ch = curl_init();
 		curl_setopt_array($ch, $opts);
 		$data = curl_exec($ch);
-		$code = (int) curl_getinfo($ch, CURLINFO_HTTP_CODE);
+		$code = curl_getinfo($ch, CURLINFO_HTTP_CODE);
 		curl_close($ch);
 
 		if ($data === false || $code < 200 || $code >= 400) {
